Alteração de para a parte do projeto

+++ Grava.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Text;
 
 public static class Grava{
-  public static void GravaCusto(float custo){
+  public static void GravaCusto(double custo){
     FileStream arquivo = new FileStream("Custo.txt",FileMode.Append,FileAccess.Write);
   
     StreamWriter sw = new StreamWriter(arquivo,Encoding.UTF8);
@@ -12,7 +12,7 @@ public static class Grava{
     sw.Close();
     arquivo.Close();
   }
-    public static void GravaLucroBruto(float lucro){
+    public static void GravaLucroBruto(double lucro){
     FileStream arquivo = new FileStream("Lucro.txt",FileMode.Append,FileAccess.Write);
 
     StreamWriter sw = new StreamWriter(arquivo,Encoding.UTF8);
